Add mapbox dir, add tests.
