+++ frontend/lib/dev/example-mapbox-page.tsx
@@ -1,8 +1,20 @@
 import React, { useContext } from "react";
 import Page from "../ui/page";
-import { MapboxCityAutocomplete } from "../forms/mapbox-autocomplete";
+import { MapboxCityAutocomplete } from "../forms/mapbox/city-autocomplete";
 import { AppContext } from "../app-context";
 import { SimpleProgressiveEnhancement } from "../ui/progressive-enhancement";
+
+const MapboxWidgets: React.FC<{}> = () => {
+  return (
+    <SimpleProgressiveEnhancement>
+      <MapboxCityAutocomplete
+        label="What city do you live in?"
+        onChange={(item) => console.log("onChange", item)}
+        onNetworkError={(err) => console.error("onNetworkError", err)}
+      />
+    </SimpleProgressiveEnhancement>
+  );
+};
 
 export const ExampleMapboxPage: React.FC<{}> = () => {
   const isMapboxEnabled = !!useContext(AppContext).server.mapboxAccessToken;
@@ -12,17 +24,7 @@ export const ExampleMapboxPage: React.FC<{}> = () => {
       <div className="content">
         <p>Mapbox integration is {isMapboxEnabled ? "enabled" : "disabled"}.</p>
       </div>
-      <SimpleProgressiveEnhancement>
-        <MapboxCityAutocomplete
-          label="What city do you live in?"
-          onChange={(item) => {
-            console.log("CHANGE", item);
-          }}
-          onNetworkError={(err) => {
-            console.error("ERROR", err);
-          }}
-        />
-      </SimpleProgressiveEnhancement>
+      {isMapboxEnabled && <MapboxWidgets />}
     </Page>
   );
 };
