+++ frontend/lib/forms/mapbox/tests/data.ts
@@ -1,0 +1,11 @@
+import _BROOKLYN from "./brooklyn.json";
+import { MapboxFeature, MapboxResults } from "../common.js";
+
+export const BROOKLYN_MAPBOX_FEATURE = _BROOKLYN as MapboxFeature;
+
+export const BROOKLYN_MAPBOX_RESULTS: MapboxResults = {
+  type: "FeatureCollection",
+  query: ["brooklyn"],
+  features: [BROOKLYN_MAPBOX_FEATURE],
+  attribution: "mapbox",
+};
